refactor(kwin-effect): extract OutlinedBorder helpers to plasmazoneseffect/borders.cpp
Pulls the four per-window border methods (removeWindowBorder,
clearAllBorders, updateWindowBorder, updateAllBorders) out of
plasmazoneseffect.cpp into a new TU. Drops the main file from 5429
to 5257 lines.

No API or behavior change. Methods remain members of PlasmaZonesEffect.
